c programing

+++ 0x00-hello_world/6-size.c
@@ -11,10 +11,10 @@ int main(void)
 	int long long d;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", sizeof(a));
-	printf("Size of a int: %lu byte(s)\n", sizeof(b));
-	printf("Size of a int long: %lu byte(s)\n", sizeof(c));
-	printf("Size of a int long long: %lu byte(s)\n", sizeof(d));
-	printf("Size of a float: %lu byte(s)\n", sizeof(f));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("Size of a int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("Size of a int long: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of a int long long: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
 	return (0);
 }
